+++ models/submission.go
@@ -25,9 +25,9 @@ type Submission struct {
 }
 
 func GetAllSubmissions(authToken string) ([]Submission, error) {
-	// if !ValidToken(authToken) {
-	// 	return []Submission{}, errors.New("Unauthorized")
-	// }
+	if !ValidToken(authToken) {
+		return []Submission{}, errors.New("Unauthorized")
+	}
 	var submissions []Submission
 	var (
 		id                   int
@@ -85,9 +85,9 @@ func GetAllSubmissions(authToken string) ([]Submission, error) {
 }
 
 func FindSubmission(submissionId int, authToken string) (interface{}, error) {
-	// if !ValidToken(authToken) {
-	// 	return []Submission{}, errors.New("Unauthorized")
-	// }
+	if !ValidToken(authToken) {
+		return []Submission{}, errors.New("Unauthorized")
+	}
 	var (
 		id                   int
 		paper_id             int
@@ -161,9 +161,9 @@ func CreateSubmissionQuery(s Submission) (int, error) {
 }
 
 func CreateSubmission(body io.Reader, authToken string) (GeneralResponse, error) {
-	// if !ValidToken(authToken) {
-	// 	return []Submission{}, errors.New("Unauthorized")
-	// }
+	if !ValidToken(authToken) {
+		return []Submission{}, errors.New("Unauthorized")
+	}
 	var s Submission
 	err := json.NewDecoder(body).Decode(&s)
 	if err != nil {
@@ -180,9 +180,9 @@ func CreateSubmission(body io.Reader, authToken string) (GeneralResponse, error)
 }
 
 func UpdateSubmission(submissionId int, body io.Reader, authToken string) (GeneralResponse, error) {
-	// if !ValidToken(authToken) {
-	// 	return []Submission{}, errors.New("Unauthorized")
-	// }
+	if !ValidToken(authToken) {
+		return []Submission{}, errors.New("Unauthorized")
+	}
 	var s Submission
 	err := json.NewDecoder(body).Decode(&s)
 	query := `
@@ -215,9 +215,9 @@ func UpdateSubmission(submissionId int, body io.Reader, authToken string) (Gener
 }
 
 func DeleteSubmission(submissionId int, authToken string) (GeneralResponse, error) {
-	// if !ValidToken(authToken) {
-	// 	return GeneralResponse{Message: "Unauthorized"}, errors.New("Unauthorized")
-	// }
+	if !ValidToken(authToken) {
+		return GeneralResponse{Message: "Unauthorized"}, errors.New("Unauthorized")
+	}
 	query := `DELETE FROM submissions WHERE id=$1`
 	res, err := database.DB.Exec(query, submissionId)
 	rowCount, err := res.RowsAffected()
